ENH: The scripts now do not offer problem with already anonymous folders.
The scripts check if nii and subjectNAME folders already exist and jump this step, whithout giving names problems in the whole procedure.
With this improvements, it is able to run the code again without need to change the folders names again.

Also, the remove files were more restrictive, which only exclude tmp* files from nii folders. The intermediate .nii.gz files continuous in the nii folder in order to check the correcting procedures to each subject.