ENH: build_template now creates the DTI probabilities maps.
The DTI FA, MD, V, L, MO and S0 images are generated as a vector image.
4D DTI maps informs the pixel probability for each voxel, using the number of subjects added in the DICOM folder.

The .bvec .bval files are copied from one individual.
It is needed finish the vector.sh script in order to correct and generate a global .bvec and .bval files which represents the DTI sample.