ENH: Created FA, MD and S0 volumes from the mean DTI and other minor adjustments.
build_template: Added a choice to clean /nii folders in order to not use so much disk space.
config/config_var.conf: Added the area to choose if you want to clean the nii folders.
postprocessing/merge_temp: Added a final processing to create the mean DTI maps with the whole subjects data. The FA, MD and S0 volumes are reconstructed.